+++ tests/testthat/test_units.R
@@ -10,9 +10,9 @@ test_that("as.unit", {
     expect_equal(as.unit(), list(unit=expression(), scale=""))
 })
 
-if (1 == length(list.files(path=".", pattern="local_data"))) {
+if (dir.exists("local_data")) {
     test_that("units in a CTD file of type WOCE (style 1)", {
-        woce <- read.ctd.woce("local_data/18HU2010014_00003_00001_ct1.csv")
+        woce <- read.ctd.woce("local_data/ctd/18HU2010014_00003_00001_ct1.csv")
         ## test units (issue 1194)
         expect_equal(woce[["pressureUnit"]], list(unit=expression(dbar), scale=""))
         expect_equal(woce[["temperatureUnit"]], list(unit=expression(degree*C), scale="IPTS-68"))